+++ mypy/partially_defined.py
